Season folder renamed for consistency between platforms.
