+++ fabric/src/main/java/club/iananderson/seasonhud/impl/minimaps/CurrentMinimap.java
@@ -5,11 +5,10 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.level.Level;
 
-
 import java.util.Objects;
 
 import static club.iananderson.seasonhud.config.Config.enableMod;
-import static club.iananderson.seasonhud.impl.fabricseasons.Calendar.calendar;
+import static club.iananderson.seasonhud.impl.seasons.Calendar.calendar;
 import static net.minecraft.world.level.Level.OVERWORLD;
