+++ forge/src/main/java/club/iananderson/seasonhud/impl/seasons/Calendar.java
@@ -1,30 +1,32 @@
-package club.iananderson.seasonhud.impl.fabricseasons;
+package club.iananderson.seasonhud.impl.seasons;
 
-import club.iananderson.seasonhud.config.Config;
-import dev.emi.trinkets.api.TrinketComponent;
-import dev.emi.trinkets.api.TrinketsApi;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
+import net.minecraftforge.fml.ModList;
+import sereneseasons.api.SSItems;
+import top.theillusivec4.curios.api.CuriosApi;
+import top.theillusivec4.curios.api.SlotResult;
 
-import java.util.Optional;
+import java.util.List;
 
-import static io.github.lucaargolo.seasonsextras.FabricSeasonsExtras.SEASON_CALENDAR_ITEM;
-import static club.iananderson.seasonhud.SeasonHUD.*;
+import static club.iananderson.seasonhud.config.Config.needCalendar;
+
+
 
 public class Calendar {
     public static boolean invCalendar;
-    public static Item calendar;
+
+    public static boolean curiosLoaded() {
+        return ModList.get().isLoaded("curios");
+    }
+
+    public static Item calendar = SSItems.CALENDAR.get();
 
     public static boolean calendar() {
-        if(extrasLoaded()){
-            calendar = SEASON_CALENDAR_ITEM;
-        }
-        else calendar = null;
-
-        if (Config.needCalendar.get() & extrasLoaded()) {
+        if (needCalendar.get()) {
             Minecraft mc = Minecraft.getInstance();
             LocalPlayer player = mc.player;
 
@@ -35,6 +37,7 @@ public class Calendar {
                 invCalendar = (slot >= 0);
 
             }
+
         }
         else invCalendar = true;
 
@@ -52,12 +55,9 @@ public class Calendar {
     }
 
     private static int findCuriosCalendar(Player player, Item item) {
-        if (curiosLoaded() && extrasLoaded()) {
-            Optional<TrinketComponent> findCalendar = TrinketsApi.getTrinketComponent(player);
-            if(findCalendar.get().isEquipped(item)){
-                return 1;
-            }
-            else return 0;
+        if (curiosLoaded()) {
+            List<SlotResult> findCalendar = CuriosApi.getCuriosHelper().findCurios(player, item);
+            return findCalendar.size();
         }
         else return 0;
     }
